pyproject.toml
[project]
name = "content-creator"
version = "0.1.0"
description = "PoC: Agencia de contenido multiagente (n8n → CrewAI) con Chainlit y trazabilidad."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.80.0",
    "chainlit>=2.0.0",
    "python-dotenv>=1.0.0",
    "openlit>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
